requirements.txt
Django==3.2
djangorestframework==3.12.2
drf-spectacular==0.19.0
djangorestframework-api-key==2.*